Move button to new file.
