+++ client/my-sites/stats/stats-module-utm/index.jsx
@@ -1,5 +1,4 @@
-import { Button, Gridicon, StatsCard } from '@automattic/components';
-import { saveAs } from 'browser-filesaver';
+import { StatsCard } from '@automattic/components';
 import classNames from 'classnames';
 import { useTranslate } from 'i18n-calypso';
 import { useState } from 'react';
@@ -14,6 +13,7 @@ import StatsModulePlaceholder from '../stats-module/placeholder';
 import StatsModuleDataQuery from '../stats-module/stats-module-data-query';
 import statsStrings from '../stats-strings';
 import UTMDropdown from './stats-module-utm-dropdown';
+import UTMExportButton from './utm-export-button';
 
 const OPTION_KEYS = {
 	SOURCE_MEDIUM: 'utm_source,utm_medium',
@@ -145,7 +145,7 @@ const StatsModuleUTM = ( { siteId, period, postId, query, summary, className } )
 					/>
 					{ showFooterWithDownloads && (
 						<div className="stats-module__footer-actions stats-module__footer-actions--summary">
-							<ExportButton data={ data } />
+							<UTMExportButton data={ data } />
 						</div>
 					) }
 				</>
@@ -153,70 +153,5 @@ const StatsModuleUTM = ( { siteId, period, postId, query, summary, className } )
 		</>
 	);
 };
-
-function ExportButton( { data } ) {
-	// Button set up.
-	const translate = useTranslate();
-	const buttonLabel = translate( 'Download data as CSV' );
-	const shouldDisableExport = data && data.length !== 0 ? false : true;
-	const shouldDrawBorder = true;
-
-	// Turns the working data into a flattened array of objects.
-	// Preserves the original data but updates the label for export.
-	const flattenDataForExport = ( originalData ) => {
-		const newData = [];
-		// Map the array of objects.
-		originalData.forEach( ( row ) => {
-			// Wrap label in quotes and push values.
-			let newLabel = `"${ row.label }"`;
-			newData.push( { label: newLabel, value: row.value } );
-			// Flatten children if present.
-			const children = row?.children;
-			if ( children ) {
-				const newChildren = children.map( ( child ) => {
-					newLabel = `"${ row.label } > ${ child.label }"`;
-					return { ...child, label: newLabel };
-				} );
-				newData.push( ...newChildren );
-			}
-		} );
-		return newData;
-	};
-
-	// Turns the flat array into a CSV string.
-	const prepareDataForDownload = ( flatData ) => {
-		const csvData = flatData
-			.map( ( row ) => {
-				return `${ row.label },${ row.value }`;
-			} )
-			.join( '\n' );
-
-		return csvData;
-	};
-
-	const initiateDownload = ( event ) => {
-		event.preventDefault();
-
-		// TODO: Provide a better default file name.
-		const fileName = 'my-data.csv';
-		const flattenedData = flattenDataForExport( data );
-		const csvData = prepareDataForDownload( flattenedData );
-		const blob = new Blob( [ csvData ], { type: 'text/csv;charset=utf-8' } );
-
-		saveAs( blob, fileName );
-	};
-
-	return (
-		<Button
-			className="stats-download-csv"
-			compact
-			onClick={ initiateDownload }
-			disabled={ shouldDisableExport }
-			borderless={ shouldDrawBorder }
-		>
-			<Gridicon icon="cloud-download" /> { buttonLabel }
-		</Button>
-	);
-}
 
 export { StatsModuleUTM as default, StatsModuleUTM, OPTION_KEYS };
